Fix percentile calculation, there was an issue with sorting
The last item of the durations array was being left out of the sort, and
sometimes I was seeing larger p75 values than p995, which is nonsensical. I've
fixed the issue and added some tests and asserts that check for sanity.

The average and standard deviation functions looked like they could use some
tweaks too.